src/app/requirements.txt
django==1.8.10
pandas
pymysql
django-bootstrap-toolkit
apscheduler
sqlalchemy
requests